feat: update option and optional value, update validation, update product,  format code, refactor some code, add comment to code, update model

+++ src/controllers/variants.controller.ts
@@ -15,22 +15,13 @@ import { AppError } from '@/utils/errorHandle';
 import { sortOptions } from '@/utils/sortOption';
 import moment from 'moment';
 // Validator schema
-import createError from 'http-errors';
 import { slugify } from '@/utils/formatters';
 
 //! Option controllers
 //Lấy tất cả các option của sản phẩm
-/**
- *
- * @param req
- * @param res
- * @param next
- * @returns
- */
 const getAllOption: RequestHandler = async (req, res, next) => {
+  const { product_id } = req.params;
   try {
-    const { product_id } = req.params;
-
     const options = await Option.find({ product_id });
 
     // Function lấy optional values
@@ -78,9 +69,8 @@ const getAllOption: RequestHandler = async (req, res, next) => {
 
 // Lấy thông tin một option cụ thể
 const getOneOption: RequestHandler = async (req, res, next) => {
+  const { product_id, option_id } = req.params;
   try {
-    const { product_id, option_id } = req.params;
-
     // Tìm option theo ID
     const option = await Option.findById(option_id).select('_id name');
 
@@ -108,31 +98,27 @@ const getOneOption: RequestHandler = async (req, res, next) => {
 };
 
 // Tạo một option mới
-//TODO: Update check duplicate
 const createOption: RequestHandler = async (req, res, next) => {
+  const { product_id } = req.params;
+  const { name } = req.body;
   try {
-    const { product_id } = req.params;
-
     // Chuẩn bị payload cho option
     const payload = {
       ...req.body,
       product_id: product_id,
     };
 
-    // // Xác thực payload
-    // const result = optionSchema.safeParse(payload);
-    // if (!result.success) {
-    //   const errors: Record<string, string> = {};
-    //   result.error.errors.forEach((e) => {
-    //     errors[e.path.join('.')] = e.message;
-    //   });
-    //   //TODO: Update sau
-    //   return res.status(StatusCodes.BAD_REQUEST).json({
-    //     message: 'Validation errors',
-    //     errors,
-    //   });
-    // }
+    const checkOption = await Option.findOne({
+      product_id: product_id,
+      name: name,
+    });
 
+    // Check if option exist, if checkOption = true, option exist
+    if (checkOption) {
+      return res.status(StatusCodes.CONFLICT).json({
+        message: messagesError.BAD_REQUEST,
+      });
+    }
     // Tạo option mới
     const doc = await Option.create(payload);
 
@@ -147,25 +133,9 @@ const createOption: RequestHandler = async (req, res, next) => {
 
 //Cập nhật thông tin một option
 const updateOption: RequestHandler = async (req, res, next) => {
+  const { option_id } = req.params;
+  const payload = req.body;
   try {
-    const { option_id } = req.params;
-    const payload = req.body;
-
-    // Xác thực dữ liệu gửi lên bằng Zod
-    // //TODO Update lại
-    // const result = optionSchema.safeParse(payload);
-    // if (!result.success) {
-    //   const errors: Record<string, string> = {};
-    //   result.error.errors.forEach((e) => {
-    //     errors[e.path.join('.')] = e.message;
-    //   });
-    //   return res.status(StatusCodes.BAD_REQUEST).json({
-    //     status: StatusCodes.BAD_REQUEST,
-    //     message: 'Validation errors',
-    //     errors,
-    //   });
-    // }
-
     // Tìm và cập nhật option
     const doc = await Option.findByIdAndUpdate(option_id, payload, {
       new: true,
@@ -189,9 +159,8 @@ const updateOption: RequestHandler = async (req, res, next) => {
 
 //Xóa một option
 const deleteOption: RequestHandler = async (req, res, next) => {
+  const { option_id } = req.params;
   try {
-    const { option_id } = req.params;
-
     // Tìm option theo ID
     const option = await Option.findById(option_id);
 
@@ -218,20 +187,27 @@ const deleteOption: RequestHandler = async (req, res, next) => {
 
 //! Optional Value Controller
 const getAllOptionalValue: RequestHandler = async (req, res, next) => {
+  const { product_id, option_id } = req.params;
   try {
-    const { product_id, option_id } = req.params;
-
-    const product = await Product.findById(product_id);
+    // Check product exist
+    const product = await Product.findById({ _id: product_id });
     if (!product) {
       return res.status(StatusCodes.NOT_FOUND).json({
         message: messagesError.NOT_FOUND,
       });
     }
 
+    // Find optional value by option id and product di
     const optionalValues = await OptionalValue.find({
       product_id,
       option_id,
     }).select('_id label value created_at updated_at');
+
+    if (!optionalValues) {
+      return res.status(StatusCodes.NOT_FOUND).json({
+        message: messagesError.NOT_FOUND,
+      });
+    }
 
     res.status(StatusCodes.OK).json({
       message: messagesSuccess.GET_OPTION_VALUE_SUCCESS,
@@ -243,9 +219,9 @@ const getAllOptionalValue: RequestHandler = async (req, res, next) => {
 };
 
 const getOneOptionalValue: RequestHandler = async (req, res, next) => {
+  const { value_id } = req.params;
   try {
-    const { value_id } = req.params;
-
+    // Find one optional value
     const optionValue = await OptionalValue.findById(value_id).select(
       '_id label value created_at updated_at'
     );
@@ -264,32 +240,55 @@ const getOneOptionalValue: RequestHandler = async (req, res, next) => {
   }
 };
 
-//TODO: Check the ref option label exist from optional models
 const createOptionalValue: RequestHandler = async (req, res, next) => {
+  /**
+   * @value_id : Optional value id
+   */
+  const { product_id, option_id } = req.params;
+  const { label, value } = req.body;
   try {
-    const { product_id, option_id } = req.params;
+    // Find option exist in product ID
+    const options = await Option.findById(option_id);
+
+    // Check option exist
+    if (!options) {
+      return res.status(StatusCodes.NOT_FOUND).json({
+        message: messagesError.NOT_FOUND,
+      });
+    }
+    const optionalValue = await OptionalValue.findOne({
+      product_id,
+      option_id,
+      label,
+    });
+
+    // Check label exist in optional value
+    if (optionalValue) {
+      throw new AppError(StatusCodes.CONFLICT, messagesError.NOT_FOUND);
+    }
+
+    // Check label in optional value and name in option was equally
+    if (label !== options.name) {
+      return res.status(StatusCodes.BAD_REQUEST).json({
+        message:
+          'The label of OptionalValue must match the label of the corresponding Option',
+      });
+    }
+
+    // Check if label and value was string type
+    if (typeof label !== 'string' || typeof value !== 'string') {
+      return res.status(StatusCodes.BAD_REQUEST).json({
+        message: 'Label and value must be strings',
+      });
+    }
+
     const payload = {
       ...req.body,
       option_id,
       product_id,
     };
 
-    // Xác thực payload với schema Zod
-    // // TODO: Update code sau
-    // const result = optionalValuesSchema.safeParse(payload);
-
-    // if (!result.success) {
-    //   const errors: Record<string, string> = {};
-    //   result.error.errors.forEach((e: any) => {
-    //     errors[e.path.join('.')] = e.message;
-    //   });
-    //   // TODO: Update
-    //   return res.status(StatusCodes.BAD_REQUEST).json({
-    //     message: messagesError.BAD_REQUEST,
-    //     errors,
-    //   });
-    // }
-
+    // Create new optional value
     const doc = await OptionalValue.create(payload);
 
     return res.status(StatusCodes.CREATED).json({
@@ -302,27 +301,13 @@ const createOptionalValue: RequestHandler = async (req, res, next) => {
 };
 
 const updateOptionalValue: RequestHandler = async (req, res, next) => {
+  /**
+   * @value_id : Optional value id
+   */
+  const { value_id } = req.params;
   try {
-    const { value_id } = req.params;
     const payload = req.body;
-
-    // Xác thực payload với schema Zod
-    // // TODO: update
-    // const result = optionalValuesSchema.safeParse(payload);
-
-    // if (!result.success) {
-    //   const errors: Record<string, string> = {};
-    //   result.error.errors.forEach((e: any) => {
-    //     errors[e.path.join('.')] = e.message;
-    //   });
-
-    //   // TODO: Update
-    //   return res.status(StatusCodes.BAD_REQUEST).json({
-    //     message: messagesError.BAD_REQUEST,
-    //     errors,
-    //   });
-    // }
-
+    // Find optional value and update
     const doc = await OptionalValue.findOneAndUpdate(
       { _id: value_id },
       { ...payload, updated_at: moment().toISOString() },
@@ -345,16 +330,19 @@ const updateOptionalValue: RequestHandler = async (req, res, next) => {
 };
 
 const deleteOptionalValue: RequestHandler = async (req, res, next) => {
+  /**
+   * @value_id : Optional value id
+   */
+  const { value_id } = req.params;
   try {
-    const { value_id } = req.params;
-
+    // Find optional value
     const doc = await OptionalValue.findById(value_id);
     if (!doc) {
       return res.status(StatusCodes.NOT_FOUND).json({
         message: messagesError.NOT_FOUND,
       });
     }
-
+    // After find , delete optional value with id
     await OptionalValue.deleteOne({ _id: value_id });
 
     return res.status(StatusCodes.OK).json({
@@ -367,18 +355,13 @@ const deleteOptionalValue: RequestHandler = async (req, res, next) => {
 };
 
 //! Variant Controllers
-// Hàm để lấy tất cả biến thể cho một sản phẩm
-/**
- *
- * @param req
- * @param res
- * @param next
- * @returns
- */
+// Get all variant exist in product
 const getAllVariant: RequestHandler = async (req, res, next) => {
+  /**
+   * @product_id : Product id
+   */
+  const { product_id } = req.params;
   try {
-    const { product_id } = req.params;
-
     // Lấy tất cả SKUs cho sản phẩm
     const SKUs = await Sku.find({ product_id });
 
@@ -389,7 +372,7 @@ const getAllVariant: RequestHandler = async (req, res, next) => {
       const value = await OptionalValue.findById(id).select('-_id value label');
       return value;
     };
-
+    // Get all color exist in product
     const getProductColor = async (array: any[]) => {
       const option = options.find(
         (opt) => opt.name === 'color' || opt.name === 'mau'
@@ -455,22 +438,38 @@ const getAllVariant: RequestHandler = async (req, res, next) => {
   }
 };
 
-// Hàm để lưu các biến thể
+// Create variant and save multiple variants
 const saveVariant: RequestHandler = async (req, res, next) => {
+  /**
+   * @product_id : Product id
+   */
+  const { product_id } = req.params;
   try {
-    const { product_id } = req.params;
-
+    // Check product exist
     const product = await Product.findById(product_id).select(
       '-_id name SKU slug price price_before_discount price_discount_percent'
     );
+    if (!product) {
+      return res.status(StatusCodes.NOT_FOUND).json({
+        status: StatusCodes.NOT_FOUND,
+        message: 'Product not found',
+      });
+    }
 
-    // Xóa tất cả SKUs và biến thể hiện có
+    // Delete all variant and SKU exist in product
     await Variant.deleteMany({ product_id });
     await Sku.deleteMany({ product_id });
 
-    // Lấy các option
+    // Check options exist
     const options = await Option.find({ product_id }).select('_id name');
+    if (options.length === 0) {
+      return res.status(StatusCodes.NOT_FOUND).json({
+        message: messagesError.NOT_FOUND,
+      });
+    }
 
+    //TODO: Move this function to utils
+    // Get optional value exist in options
     const getOptionValues = async (option: OptionType, id: Types.ObjectId) => {
       const optionValues = await OptionalValue.find({ option_id: id }).select(
         '_id label value'
@@ -478,10 +477,14 @@ const saveVariant: RequestHandler = async (req, res, next) => {
       return { ...option, option_values: optionValues };
     };
 
+    // Create option values base on options
     const docs = await Promise.all(
       options.map((option) => getOptionValues(option.toObject(), option._id))
     );
+    console.log(docs);
 
+    //TODO: Move this function to utils
+    // Generate variants
     const generateVariant = (input: any[]) => {
       if (input.length === 0) return [];
 
@@ -517,13 +520,12 @@ const saveVariant: RequestHandler = async (req, res, next) => {
       return result;
     };
 
-    if (!product) {
-      return res.status(StatusCodes.NOT_FOUND).json({
-        status: StatusCodes.NOT_FOUND,
-        message: 'Product not found',
-      });
-    }
+    // Generate variants
+    //TODO: check error
     const variants = generateVariant(docs);
+
+    // Create array of SKUs from variants
+    //TODO: Check error
     const arraySKUs = Array(variants.length).fill({
       ...product.toObject(),
       product_id,
@@ -531,6 +533,8 @@ const saveVariant: RequestHandler = async (req, res, next) => {
       assets: [],
     });
 
+    // Create SKU base on variants
+    //TODO: Check error
     const SKUs = await Promise.all(
       arraySKUs.map(async (item, index) => {
         const variantValues = variants[index].map((v) => v.label).join('-');
@@ -547,6 +551,8 @@ const saveVariant: RequestHandler = async (req, res, next) => {
       })
     );
 
+    //TODO: Move this function to utils
+    // Create variant options if there was have options and optional value
     const variantOptions = (variants: any[], SKUs: any[]) => {
       let result: any[] = [];
       for (let index in SKUs) {
@@ -566,11 +572,14 @@ const saveVariant: RequestHandler = async (req, res, next) => {
       return result;
     };
 
+    // Create variant option with variant and SKUs
     const data = variantOptions(variants, SKUs);
+    //TODO: Check error
+    // Create variant data with SKU
     const createVariantData = await Promise.all(
       data.map((item) => Variant.create(item))
     );
-
+    //TODO: Check error
     return res.status(StatusCodes.CREATED).json({
       message: messagesSuccess.CREATED,
       res: createVariantData,
@@ -580,18 +589,22 @@ const saveVariant: RequestHandler = async (req, res, next) => {
   }
 };
 
-// Hàm để xóa một biến thể
+// Delete onr variant
 const deleteVariant: RequestHandler = async (req, res, next) => {
+  /**
+   * @product_id : Product id
+   * @ski_id : SKU id
+   */
+  const { product_id, sku_id } = req.params;
   try {
-    const { product_id, sku_id } = req.params;
+    // Check if SKU exist
     const sku = await Sku.findById(sku_id);
-
     if (!sku) {
       throw new AppError(StatusCodes.NOT_ACCEPTABLE, messagesError.NOT_FOUND);
     }
 
+    // Find variant
     const variants = await Variant.find({ sku_id });
-
     // Đánh dấu SKU là đã xóa
     sku.deleted_at = new Date();
     await sku.save();
@@ -613,11 +626,14 @@ const deleteVariant: RequestHandler = async (req, res, next) => {
   }
 };
 
-// Hàm để lấy một biến thể
+// Get one variant
 const getOneVariant: RequestHandler = async (req, res, next) => {
+  /**
+   * @product_id : Product id
+   * @ski_id : SKU id
+   */
+  const { product_id, sku_id } = req.params;
   try {
-    const { product_id, sku_id } = req.params;
-
     const sku = await Sku.findOne({ _id: sku_id }).select(
       '-deleted -deleted_at -created_at -updated_at'
     );
@@ -665,25 +681,14 @@ const getOneVariant: RequestHandler = async (req, res, next) => {
   }
 };
 
-// Hàm để cập nhật một biến thể
+// Update one variant
 const updateVariant: RequestHandler = async (req, res, next) => {
+  /**
+   * @ski_id : SKU id
+   */
+  const { sku_id } = req.params;
   try {
-    const { sku_id } = req.params;
     const body = req.body;
-
-    // Xác thực dữ liệu đầu vào
-    // //TODO: Update
-    // const result = variantSchema.safeParse(body);
-    // if (!result.success) {
-    //   const errors: Record<string, string> = {};
-    //   result.error.errors.forEach((e) => {
-    //     errors[e.path.join('.')] = e.message;
-    //   });
-    //   return res.status(StatusCodes.BAD_REQUEST).json({
-    //     message: 'Validation errors',
-    //     errors,
-    //   });
-    // }
 
     const { options, ...payload } = body;
 
@@ -736,22 +741,22 @@ const updateVariant: RequestHandler = async (req, res, next) => {
   }
 };
 export {
-  createOption,
-  createOptionalValue,
-  deleteOption,
-  deleteOptionalValue,
-  deleteVariant,
   // Option Properties
+  getOneOption,
   getAllOption,
+  deleteOption,
+  createOption,
+  getOneOptionalValue,
   // Optional Value
+  deleteOptionalValue,
+  createOptionalValue,
+  updateOptionalValue,
   getAllOptionalValue,
   // Variant
   getAllVariant,
-  getOneOption,
-  getOneOptionalValue,
+  deleteVariant,
   getOneVariant,
   saveVariant,
   updateOption,
-  updateOptionalValue,
   updateVariant,
 };
